requirements.txt
# Core ML & Audio
openai-whisper==20231117
torch>=2.2.0,<2.9.0
torchaudio>=2.2.0,<2.9.0
numpy>=1.24.0

# Audio Processing
ffmpeg-python>=0.2.0

# Speaker Diarization Pipeline
speechbrain>=0.5.16
scikit-learn>=1.3.0

# VAD (Silero) - loaded via torch.hub
requests>=2.31.0
tqdm>=4.66.0

# Utilities (ignored on Python >= 3.4)
pathlib2>=2.3.7.post1; python_version < "3.4"